clean up
All files are now name *-system.morph or *-stratum.morph, to
make it clear what they are. This is redundant, but when we have
both a "linux" chunk and a "linux" stratum, it's good to be clear.

Also, rename core to foundation.

Also, fix references to be the proper branches, for bootstrapping.
Hopefully, the bootstrap branches will work for real building as well,
but if not, we'll deal with that later.