bootstrap-pass2: Remove unnecessary chunks?
A lot of user tools are built that shouldn't be necessary

some of util-linux will be needed to make system images, but
currently we build the tools before we try to make a system image

psmisc provides process information, this should not be needed
to build anything and busybox provides its functionality

e2fsprogs is needed to make ext filesystems, but this should not
be needed to build chunks, it may be required for system images

iana-etc provides some internet protocol information, this should
not be needed for building

procps, see psmisc

rsync was needed to download translation files, it should
not be needed, if it is needed then it should be fixed not to

wget was needed to download during build, we should not download
during a build. This required gnutls which required libnettle.
gmp should not be needed now

vim should never be needed to build, it may be built in pass 3

sudo has never been configured correctly and is not needed for
building software

busybox-devel has its important programs provided by coreutils etc.
Some coreutils are not substitutable for busybox when building,
ideally busybox should be used for everything else

linux should not be needed for compilation