bootstrap: add linux-user-chroot
It builds just like autotools, so its morphology can be inferred
for pass 2 and devel, but pass 1 needs some configuration because
the headers needed to build it aren't in debian squeeze.

autogen.sh uses `which` to test whether autoreconf exists, which
we only have in busybox, which wasn't built in pass 1. It is built
later in pass 2, so it is easier to build linux-user-chroot later.

Also remove sudo from bootstrap, we don't use it.