Switch to the stable baserock/3.2-headers branch of linux-api-headers.
This is done so that when we move from one version of the linux kernel
to another (e.g. by making a single change), we don't have to rebuild
the headers all the time (which so far happened because linux.morph and
linux-api-headers.morph were located in the same branch).