Add {imx53,x86_64-generic}-bootstrap-pass3 strata.
These are used to group together what we want the results of the
third pass of the bootstrap to be. Building one of these strata
in a staging chroot will make the stuff of e.g. devel available to
stuff that needs to be compiled for the BSP.

Before I change baserock-bootstrap to build these strata in the
third pass, I'll see if they actually solve our problem of not sharing
the build results of devel with the builds needed for e.g. a BSP (in
our case this was nasm having been installed into the devel build
chroot and then not being available in the separate BSP chroot).