Initial attempt at adding Vala to devel
Not tested yet.

Note that there is a vala-bootstrap module in gnome git which has
the same purpose as our baserock/bootstrap branch. However, since
we will need to do the same procedure for other languages in the
future, perhaps it's better to be consistent for now, we can
easily drop our branch in favour of upstream's in the future.