Reorganise morphologies for build-essential
This patch rewrites the Baserock morphologies so that they can be
built using a minimal "build-essential" staging filler, as described
in my previous patch mail.

The Baserock system is split up into three strata:

  - build-essential: the minimum set of tools that can rebuild
    itself (from tarballs)
  - core: the remaining set of components necessary to produce
    a Baserock system which can rebuild itself from source
  - tools: auxiliary development tools and libraries

The 'build-essential' stratum does not specify any build
dependencies, because it expects to be built with a staging filler
containing itself. In effect, the stratum has a circular dependency
on itself, which we solve by using a staging filler. The 'core'
stratum depends on only on 'build-essential'. This ensures that all
build dependencies between chunks in 'core' are explicit. These
strata can also be built with the staging filler from the water-bomb
release.

The distinction between a base and a devel system is no longer
defined by the 'foundation' and 'devel' strata (in practice these
had various hidden build dependencies between each other). The size
of a base system is larger than before, due to more development tools
being in 'core'.