Split 'foundation' stratum back out of 'core'
Mixing together the set of build tools required to build the base
system and the components in the base system itself has one big issue:
changing any component of the base system triggers a rebuild of every
higher stratum, which even for simple Baserock systems means
constantly rebuilding Linux.

To solve this, I have turned 'core' into a sort of larger version of
build-essential (I am considering renaming it to 'build-core' or even
just 'build'). Chunks which are required for running a base system
but not during the build of other components are now in the
resurrected 'foundation' stratum, along with their specific
dependencies (which stratum a chunk should go in is roughly decided
by its utility to chunks outside of foundation).