Add glanceclient Stratum morphology.
glanceclient.morph: Stratum morphology to add glanceclient. This stratum
also adds all the necessary dependencies needed.

Glanceclient is needed to provide the development system the capacity
to deploy images into an openstack system.
This is a client library for Glance built on the OpenStack Images API.
It provides a Python API (the glanceclient module) and a command-line
tool (glance).