Move pyyaml from 'tools' to 'core'.
Also moved python-setuptools because is a dependency of pyyaml. Pyyaml
is moved to core because is a dependency of 'cloudinit-support' stratum
and moving it give us the capacity of buid base systems with cloduinit.

core.morph: added pyyaml and python-setuptools
tools.morph: deleted pyyaml and python-setuptools
cross-bootstrap.morph: deleted pyyaml and python-setuptools