Make Trove deployment handle new Lorry Controller
Add install-files to the list of configuration extensions, on general
principles. It costs nothing to have it, and is a nice way to install,
say, ssh keys.

In trove.configure, support the optional LORRY_CONTROLLER_MINIONS
configuration key to decide how many Lorry Controller MINION instances
may run at the same time. It has a default in case it is not set.

Create /etc/lorry-controller/webapp.conf. This needs to have some
information from the deployment, so it can't be just installed by
the Lorry Controller chunk.