trove: Do subsitutions in /etc at deploy-time, not at first boot
This makes it easier to merge changes between /etc when doing upgrades.

Lorry and Mason config goes into local-config.git, not in /etc, so this
still needs to be done at first boot when we can run Gitano commands on
the target. For this reason the /etc/trove-setup.sed file is still
stored in the resulting system.