Allow minimal systems to be bootable
This adds a busybox-init.configure extension, which if
INIT_SYSTEM=busybox is set, will configure it to boot with busybox.

The existing minimal system has been changed to use the busybox-init
configuration extension.

minimal-system-deploy.morph has been added as an example, showing that
INIT_SYSTEM: busybox needs to be set.