Build cpython with readline and ncurses
At the moment if you try to use the arrow keys in a python shell
then you just get some unhelpful output.

This builds cpython with readline so the arrow keys can do useful things

It also adds ncurses as a build dependency, so that the 'curses' module
can be imported.