Add example systems which use an initramfs
initramfs-test.morph is embedded in the initramfs.write help, so may be
removed.

The initramfs-x86_64 system is the minimal system with an extra chunk
added.

The initramfs-utils stratum is required to contain _just_ that chunk,
but the "minimal" system is much larger than I would like our initramfs
to be (i.e. a statically linked binary using musl), so in future the
initramfs-utils stratum may have more initramfs specific content.