Prevent distbuild nodes from filling up their disks
Distbuild workers run `morph gc` before running a build, to remove stale
artifacts, but it turns out that they were configured in such a way that
the command didn't do anything. Because of this, it was possible for
distbuild nodes to break due to filling up their disks with cached git
repos and artifacts.