Explicitly separate linux chunk names for bsp variants
Simplifying 'morph edit' syntax showed that we have several
different versions of the chunk called 'linux'. This patch
versions of linux to match what they are for.

This patch was originally written by Paul Sherwood, but the name of
the morphology also has to be set in the chunk morphology, so wasn't
sufficient by itself.

Instead of making the changes in the 7 kernel trees that we had, they
have now all been merged into one, hence the sha1s and refs have also
been updated.