Change backup-gitlab to use database dump instead of Rake script
The Gitlab Rake script for backing up doesn't backup the database
that Gitlab CI uses, and additonally it does a git bundle for each
git repository. Our GitLab instance has hundreds of git repositories,
making the bundling be a very expensive things.

Instead, we use the Postgres tool to dump all databases to a file,
and arrange to backup the dump files, the git repositories, and all
other relevant files directly.