Add new scripts for building, uploading release
These scripts are a rewrite of scripts/do-release.py and
scripts/distbuild-cluster. The biggest difference is that they split
the tasks of building the things that are to be released, and uploading
them to git.baserock.org / download.baserock.org, where do-release.py
combines both (and distbuild-cluster only builds chunk/stratum/system
artifacts, not the release images). The new scripts are also configurable
using command line options or a configuration file rather than requiring
editing of the source.

These changes will allow, for example, a CI job that builds a release,
but doesn't upload it to download.baserock.org.

The new scripts are coupled with a change to the release process, which
will be documented as a change to the release process page on
wiki.baserock.org.

The 14.29 release of Baserock was done with slightly different versions
of these scripts to make it feasible to upload things over multiple
network connections.