Allow release-upload to be configured not to upload release images
For continuous artifact cache population, we don't care so much about
the large disk images that we make available at release time.

This patch allows omitting any of the configuration required to upload
the release images to mean that we didn't want to upload them, and
continue without doing so.