Allow release scripts to be configured to only handle a subset of architectures
We currently build all architectures at once during the release process,
however for our CD pipeline we operate with one CD pipeline per
architecture.

This is not just useful for the CD pipeline work though, as it allows
one organisation to handle releases for x86, where the infrastructure
may be located in the cloud, and one organisation to handle ARM systems,
which may be located in an office.