mason: Don't checkout repo using .git suffix
Failing to do this means that the deployment uses the wrong morphology,
because build will end up using the repo without the .git suffix, so it
will never update the cached version of the repo without it.

The version with the .git suffix is only updated on the initial
checkout, but is used by deploy, so it would pick up obsolete
morphologies and not include new changes.

Rubber-stamped-by: Richard Maw