Split PCRE out into its own stratum
We need PCRE in multiple places, so rather than having multiple
definitions, and potentially differing versions, let's use the same one
from its own stratum.

The need to add the stratum entry into the systems is annoying, and
could probably do with a way of specifying that a stratum includes
another, rather than just depends on it.

Also, it seems a bit weird to have a stratum with only one chunk in it,
but I can't think of a better way.