Patch m4 to be glibc-2.20 compatible
m4 has a hacked up stdio.h which will give warnings when certain
functions are used. One of these warnings was that 'gets' is a
deprecated function. The patch removes the code that adds the
warning.

Presumably the new version of glibc finally got rid of it, which makes
it puzzling that the deprecation warning code was failing to handle the
fact that the deprecated function could have been removed.