Allow distbuild.configure to be enabled without requiring configuration
Previously if you enabled distbuild.configure but didn't set all of the
required variables, your deployment would fail. This has become much
more annoying with the introduction of the build-system family of
systems, which are indented to be suitable either as a local builder or
part of a distbuild network.