Generate .tarball-version in bison
The version number of bison appends a part of the sha1
of the commit which is being built. This was causing
failures in the autoconf when building gstreamer given
that it tries to compare the version number of bison.

This is a *workaround* to generate a clean version number
of bison.