Add Lorry and dependencies to devel systems
Lorry is used by the Baserock Import tool for fetching source code of
components that are to be imported.

It's also generally useful in devel systems as it can be used for
testing .lorry files prior to pushing them to a Trove.

Additionally, this means that devel systems now contain 'hg', 'bzr',
'svn' and 'cvs', any of which may come in handy.

This increases the size of the devel-system-x86_64-chroot system from
1.4GB to 1.5GB.