Fix Mason deployments for non-openstack-testing Mason
In mason.configure:

It's not needed to create a separate os.conf file at this point.
In this file we were puting OpenStack credentials used to create
an os.conf file with Ansible. This file was only created when
TEST_INFRASTRUCTURE_TYPE was 'openstack', and Ansible was expecting
it always.

This patch moves the OpenStack credentials to mason.conf, so Ansible
only have to read the variables from one file.

In mason.sh:

The script was always loading /etc/os.conf. This file is only created
when TEST_INFRASTRUCTURE_TYPE is 'openstack'. This patch checks that
the file exists before loading it.

In mason.conf template for Ansible.

OPENSTACK_NETWORK_ID is only present when TEST_INFRASTRUCTURE_TYPE
is 'openstack'. This patch adds a conditon in the template to
skip this value if it doesn't exist.