Use patched version of pip
This patch moves us onto a patched version of pip that allows the
import tool to obtain runtime dependencies for a python package.

We will discuss the addition of this functionality with the upstream
pip developers at a later date, we should also hopefully be able to
find a nicer way of implementing it.