openid_provider: Disable SELinux and enable persistant Journal
Default SELinux policy seems to prevent the uWSGI process from reading
stuff in /srv and raises a bunch of errors for Cherokee trying to
connect() to its socket, too. It'd be possible to fix this but I'd
rather not bother, and since we don't use SELinux in Baserock it's not
super important anyway.

If you did want to fix this I'd suggest installing 'setroubleshoot' and
reading <https://fedoraproject.org/wiki/SELinux_FAQ> as a starting
point.