frontend: Enable logging for HAProxy
Note that by default the systemd journal doesn't have any limit on the
amount of disk space it uses, and this system uses the default
configuration so it will eventually fill its own disk up with logs.

Switching to Ansible for provisioning this system should make it easier
to fix this.