Try to work around corrupted OpenStack images in a different way
I continue to find that the OpenStack images deployed by Packer are
often corrupt, in that files added during provisioning exist but are
truncated to 0 bytes.

I found <https://github.com/mitchellh/packer/issues/1462> which suggests
the problem might be that Packer takes a snapshot without shutting down
or even running 'sync' in the image. Adding lots of 'sync' calls at the
end of provisioning does seem to have helped, so far!