Work around a Packer bug when deploying development images
Issue is: <https://github.com/mitchellh/packer/issues/1656>. The problem
only exists in Red Hat-based operating systems, because for some reason
'cp' gets aliased to 'cp -i', and for some reason Packer doesn't clean
up the temporary shell scripts it creates in the system or use 'cp -f'
to overwrite them.

As a workaround, use '/bin/sh' instead of '/bin/bash' as the shell for
Docker deployments, which I think is an improvement in general because
the less clever shell code involved in deploying a system the better.