Add the ability to deploy installer systems and add examples.
The installer-x86_64 system is a system that can be used to install other
systems in an storage device. This system is intended to be booted by usb,
pxeboot... to install a Baserock system in your local disk.

The installer system requires the installer.configure extension to generate
a configuration file located in /etc/install.conf. With this extension
you can specify the following variables in a cluster morphology:

- INSTALLER_TARGET_STORAGE_DEVICE: Target storage device to install the
    Baserock system.
- INSTALLER_ROOTFS_TO_INSTALL: The location of the root filesystem that
    is going to be installed.
- INSTALLER_POST_INSTALL_COMMAND: Commands that will be run after the
    installation finishes. It defaults to `reboot -f`.

The installer-utils stratum is required to contain the installer-scripts
chunk. This chunk contains the installer script that is going to be
installed in /usr/lib/installer/installer.py

The clusters/installer-build-system-x86_64.morph file defines the deployment
of a installer system as a rawdisk image. This installer system will
install a build-system-x86_64 located in /rootfs into the /dev/sda device.
Also this cluster defines a subsystem which is the build-system that
is going to end up in /rootfs on the installer system.