Temporary Networkd fix to configure DHCP also in eth* interfaces
We assumed that the ethernet interfaces always were going to be
called en* (e.g. ens3) when using the new version of systemd (v217),
but sometimes these interfaces can be called eth* (e.g. eth0). If
this happens the current configuration is not enough.

This patch is to enable DHCP also in eth* interfaces.