Add chunks for Java, ANT and ZooKeeper
Java is sourced from the binary Java release from Oracle.  This
chunk was originally written by Francisco Marchena.

ANT is a Java build system and is needed by ZooKeeper.

ZooKeeper itself is documented at http://zookeeper.apache.org/

This patch also brings in a zookeeper test program in a seperate strata
that can be safely discarded if not required. this test program
was written by me, <mike.smith@codethink.co.uk> and is not designed to
be used in any practical way, but to showcase the functionality
of zookeeper within baserock

The ZooKeeper demonstration server and client are currently hosted
on baserock/test.

The Java binary chunk only works for x86_64.  As such, these
systems are limited to that architecture.