pxeboot.check: Multiple complete_matches is not an error
The code was assuming that if there are more than one complete_match
we were setting more options that needed and this is not always true.

For example when setting 'PXEBOOT_DEPLOYER_INTERFACE' and 'PXEBOOT_VLAN',
there will be more than one complete_matches, given that these options match
the spawn-vlan and the spawn-novlan modes.

My workaround consists on report warning messages instead of failing when
this happens.

The reason that I haven't removed the code is that I think that it could
be fixable. For now this code will report "warnings" instead of failing.