openid_provider: Get ready for production!
This includes some tweaks to how deployment is done. There is now an
'instance configuration' stage, where an Ansible playbook injects
instance-specific info like the database password into the machine, and
runs database migrations. It makes sense for this to be a separate stage
because it may need to be re-executed during the lifetime of the
instance, for example if the database passwords change.