database: Get ready for production!
To enable backups, the contents of /var/lib/mysql are now stored in a
logical volume managed by LVM. We can take a snapshot of this volume in
a matter of seconds, meaning we can take a clean, local snapshot of the
state of the database with only small amount of downtime. The snapshot
can then be mounted and its contents copied out of the system while the
MariaDB service is running again.