Tweak README
Use $keyname as a placeholder for key pair name so that I can copy and
paste the command as-is and set keypair= in my environment, instead of
manually replacing the <key-name> placeholder each time.

Also note that database/instance-mariadb-config.yml can be rerun any
time to update the MariaDB users and tables.