Upgrade systemd to v218 to get bug fixes
Trove-setup uses `systemctl enable` and it wasn't working for
template units (in this case the minion services). With this
upgrade we fix the Trove deployments.

Also some users have found some problems with the 'systemd-journald'
service, which starts logging before the system mounts the '/var'
subvolume. This upgrade is supposed to fix that as well.

Systemd v218 had an important bug:

  ~ # ssh localhost
  Connection closed by ::1

Which was dropping this error in the systemd journal:

  Assertion 'canonical' failed at src/nss-myhostname/nss-myhostname.c:204, function fill_in_hostent(). Aborting.

This bug was fixed upstream in the commit 3fdcecc87eb381ef300719e419d5863dd8a64a97
and it's also been included in this upgrade.