openid_provider: Require users to verify their email addresses
This was discussed in #baserock on 2015-01-30. We decided that since
email address is a part of a person's online identity, it is worth the
extra hassle of requiring users to verify their email address, to make
it harder for people to create a Baserock OpenID account that impersonates
another person.

Users cannot log in until they have activated their accounts by clicking
a link that's emailed to them. And they cannot use their OpenID unless
they are logged in.