Regenerate config.{guess,sub} scripts
Some packages have old config.guess and config.sub scripts which don't
recognize the aarch64-linux-gnu GNU triplet.  For some cases, it is
enough to do a `autoreconfig -if` to update those scripts.  However,
this approach cannot be used because when the configure scripts are not
compatible with the current automake, autoheader, aclocal and friends;
or when the project doesn't use automake at all (it has Makefile.in
files, but not Makefile.am files).  For those cases, we copy the system
config.guess and config.sub scripts installed by automake.