Create xstatic stratum
XStatic is a packaging standard to package external static files
as a python package, so they are easily usable on all OSes.
This stratum also adds some xstatic libraries.
This conflicts with the 'everything from Git' policy of
Baserock,but it is required right now for the Openstack system to work.