sdk.write: Fix with modern versions of 'find'
The '-perm +111' syntax has been deprecated since 2005 and has been
removed from modern versions of GNU Find.

Thus, the SDK installer script fails on Fedora 21 (GNU Find 4.5.11)
with the following error:

    find: invalid mode +111

The recommended replacement is '-perm /111' but this isn't supported by
the version of Busybox 'find' present in current Baserock 'build'
systems.

This test is being done purely to find files that are executable, and
for this purpose '-perm +u=x' is equivalent (but supported by both
versions of 'find').