pxeboot: disable preserving extended attributes
Running rsync with the '-X' option causes the following error to occur:

    ERROR: pxeboot.write failed with code 1: rsync: get_xattr_names:
llistxattr(""/src/tmp/deployments/tmpcMabnt/tmp0jCu5s"",1024) failed:
Function not implemented (38)

This requires more investigation to understand why it is failing (the
attr devel libraries were installed).  AFAIK we are not using extended
attributes in the rootfs that we create, so disable this option for the
time being.