Make stage2-linux-api-headers and stage2-gawk dependencies explicit
As discussed on the ML at [1] morph somehow infers that
stage2-linux-api-headers is required for other stage2 components.
This patch makes the inference explicit, which is more consistent,
and should not affect any current builds or cache calculations.

Also, it seems m4 requires stage2-gawk, so make that explicit too.

[1] http://listmaster.pepperfish.net/pipermail/baserock-dev-baserock.org/2015-March/012425.html